pyproject.toml
[project]
name = "medguard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.30.0",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
]
